pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-md-validator"
version = "0.1.0"
description = "Validate agent-md-specs configuration files"
readme = "README.md"
license = {text = "CC0-1.0"}
requires-python = ">=3.10"
authors = [
    {name = "TotalMarkdown.ai"},
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
agent-md-validate = "agent_md_validator.cli:main"

[project.urls]
Homepage = "https://totalmarkdown.ai"
Repository = "https://github.com/totalmarkdown/agent-md-validator"
Specifications = "https://github.com/totalmarkdown/agent-md-specs"
